% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{askUser}
\alias{askUser}
\title{askUser}
\usage{
askUser(query)
}
\arguments{
\item{query}{What do you want to ask the user?}
}
\value{
user input
}
\description{
Helper function to consistently get user input
}
\seealso{
Other utility: 
\code{\link{concatEnsembleTimeseries}()},
\code{\link{convertAD2BP}()},
\code{\link{convertBP2AD}()},
\code{\link{gaussianize}()},
\code{\link{generateEnsembleFromUncertainty}()},
\code{\link{getOs}()},
\code{\link{heuristicUnits}()},
\code{\link{loadRemote}()},
\code{\link{pullInstance}()},
\code{\link{simulateAutoCorrelatedUncertainty}()},
\code{\link{surrogateDataFun}()}
}
\concept{utility}
