% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens-package.R
\docType{package}
\name{ens-package}
\alias{ens}
\alias{ens-package}
\title{ens: Ensemble Methods for Time-Uncertain Data}
\description{
Ensemble methods and calculations for time-uncertain paleogeoscientific data, including binning, correlation, regression, principal component analysis, and spectral analysis across age ensembles, plus utilities for extracting and mapping ensembles in LiPD (Linked Paleo Data) objects. Split out of the 'geoChronR' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickmckay/ens}
  \item Report bugs at \url{https://github.com/nickmckay/ens/issues}
}

}
\author{
\strong{Maintainer}: Nicholas McKay \email{nick@nau.edu} (\href{https://orcid.org/0000-0003-3598-5113}{ORCID})

Authors:
\itemize{
  \item Julien Emile-Geay (\href{https://orcid.org/0000-0001-5920-4751}{ORCID})
  \item Deborah Khider (\href{https://orcid.org/0000-0001-7501-8430}{ORCID})
}

}
\keyword{internal}
