% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null-hypothesis.R
\name{kdePval}
\alias{kdePval}
\title{Estimate a p-value from a null distribution using kernel density estimation}
\usage{
kdePval(nulls, real, h = NA, xmin = NA, xmax = NA)
}
\arguments{
\item{nulls}{a vector of null hypothesis results}

\item{real}{a single value for the real data}

\item{h}{smoothing factor}

\item{xmin}{minimum value for the distribution}

\item{xmax}{maximum value for the distribution}
}
\value{
a list with the p-value and the kde data (x, y)
}
\description{
Estimate a p-value from a null distribution using kernel density estimation
}
\seealso{
Other ensemble: 
\code{\link{propagateUncertainty}()},
\code{\link{testNullHypothesis}()}
}
\concept{ensemble}
